# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Pokerbots Competition API
  version: "1.0.0"
  description: |
    Student-facing and admin API for the Berkeley Pokerbots competition.

    * Base path: `/api/v1`
    * Auth: `X-Api-Key: <team_api_key>` **or** `Authorization: Bearer <team_api_key>`.
      Admin routes accept `X-Admin-Key: <admin_key>` or the key of a team flagged `is_admin`.
    * Errors always use the envelope `{ "error": { "code", "message", "details" } }` with codes
      `unauthorized`, `forbidden`, `not_found`, `validation_error`, `conflict`, `rate_limited`,
      `payload_too_large`, `not_ready`, `internal_error`.
    * Bot protocol: see `docs/BOT_PROTOCOL.md`.
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
    Bearer:
      type: http
      scheme: bearer
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    TournamentConfig:
      type: object
      description: Engine configuration (see SPEC.md). All fields optional when patching.
      properties:
        table_size: { type: integer, default: 9 }
        break_threshold: { type: integer, default: 7 }
        min_table_size: { type: integer, default: 6 }
        starting_stack: { type: integer, default: 1000 }
        blinds:
          type: object
          description: "{type: geometric, starting_small, starting_big, multiplier, ante_starts_at_level?, ante_fraction_of_bb?} or {type: levels, levels: [...]}"
        level_advance:
          type: object
          properties:
            max_level_duration_secs: { type: integer, nullable: true }
            hands_per_level: { type: integer, nullable: true, description: "advance when the average hands per table in the level reaches this" }
            avg_stack_thresholds: { type: array, items: { type: array, items: { type: integer }, minItems: 2, maxItems: 2 } }
        action_timeout_ms: { type: integer, default: 1000 }
        series_length: { type: integer, default: 100 }
        rng_seed: { type: integer }
        max_hands_per_table: { type: integer, nullable: true }
    RunProgress:
      type: object
      properties:
        tournaments_total: { type: integer }
        tournaments_completed: { type: integer }
        tournaments_failed_or_cancelled: { type: integer }
    Run:
      type: object
      properties:
        run_id: { type: string, example: run_9f8e7d6c5b4a }
        kind: { type: string, enum: [nightly, ondemand] }
        status: { type: string, enum: [queued, running, finalizing, completed, failed, cancelled] }
        config:
          type: object
          properties:
            series_length: { type: integer }
            table_size: { type: integer }
            seed: { type: integer }
            action_timeout_ms: { type: integer }
            starting_stack: { type: integer }
            record_hands: { type: boolean }
            tournament: { $ref: '#/components/schemas/TournamentConfig' }
        participant_count: { type: integer }
        created_by: { type: string, description: "team id, 'admin' or 'scheduler'" }
        created_at: { type: string, format: date-time }
        snapshot_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
        error: { type: string, nullable: true }
        nightly_date: { type: string, nullable: true }
        progress: { $ref: '#/components/schemas/RunProgress' }
        participants:
          type: array
          description: Only on GET /runs/{run_id}
          items:
            type: object
            properties:
              team_id: { type: string }
              team_name: { type: string }
              submission_id: { type: string }
    TournamentSummary:
      type: object
      properties:
        index: { type: integer }
        seed: { type: integer }
        status: { type: string, enum: [queued, running, completed, failed, cancelled] }
        started_at: { type: string, nullable: true }
        finished_at: { type: string, nullable: true }
        worker_id: { type: string, nullable: true }
        total_hands: { type: integer, nullable: true }
        winner_team_id: { type: string, nullable: true }
        winner_team_name: { type: string, nullable: true }
        error: { type: string, nullable: true }
    TournamentDetail:
      allOf:
        - $ref: '#/components/schemas/TournamentSummary'
        - type: object
          properties:
            placements:
              type: array
              items:
                type: object
                properties:
                  team_id: { type: string }
                  team_name: { type: string, nullable: true }
                  place: { type: integer, description: "1 = winner; ties share a place" }
                  hands_played: { type: integer }
            eliminations: { type: array, items: { type: object } }
            levels_reached: { type: integer, nullable: true }
            duration_ms: { type: integer, nullable: true }
            spawn_failures: { type: array, nullable: true, items: { type: object } }
            hand_log_url: { type: string, nullable: true, description: "Present when hand histories were recorded (admin only)" }
...
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    NotFound:
      description: Not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Conflict
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Validation:
      description: Validation error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
security:
  - ApiKey: []
  - Bearer: []
paths:
...
  /runs:
    get:
      tags: [runs]
      security: []
      summary: List runs
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: kind, in: query, schema: { type: string, enum: [nightly, ondemand] } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
        - { name: cursor, in: query, schema: { type: string }, description: created_at cursor }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs: { type: array, items: { $ref: '#/components/schemas/Run' } }
                  next_cursor: { type: string, nullable: true }
    post:
      tags: [runs]
      summary: Enqueue an on-demand run using the current active-bot snapshot
      description: Teams are rate limited (`ondemand_runs_per_team_per_day`, `ondemand_max_series_length`); admins are not.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                kind: { type: string, enum: [ondemand, nightly], default: ondemand, description: nightly requires admin }
                mode: { type: string, enum: [series, single], default: series }
                series_length: { type: integer }
                include_inactive: { type: boolean, default: false, description: must be false }
                seed: { type: integer, nullable: true }
      responses:
        "202": { description: Queued, content: { application/json: { schema: { $ref: '#/components/schemas/Run' } } } }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "422": { $ref: '#/components/responses/Validation' }
        "429":
          description: Rate limited
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /runs/{run_id}:
    get:
      tags: [runs]
      security: []
      summary: Run detail with progress and participants
      parameters: [{ name: run_id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Run' } } } }
        "404": { $ref: '#/components/responses/NotFound' }
  /runs/{run_id}/cancel:
    post:
      tags: [runs, admin]
      security: [{ AdminKey: [] }, { ApiKey: [] }]
      summary: Cancel a run (admin)
      parameters: [{ name: run_id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Cancelled, content: { application/json: { schema: { $ref: '#/components/schemas/Run' } } } }
        "409": { $ref: '#/components/responses/Conflict' }
  /runs/{run_id}/tournaments:
    get:
      tags: [runs]
      security: []
      summary: Per-tournament summary
      parameters: [{ name: run_id, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id: { type: string }
                  tournaments: { type: array, items: { $ref: '#/components/schemas/TournamentSummary' } }
  /runs/{run_id}/tournaments/{index}:
    get:
      tags: [runs]
      security: []
      summary: Tournament detail with placements
      parameters:
        - { name: run_id, in: path, required: true, schema: { type: string } }
        - { name: index, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TournamentDetail' } } } }
        "404": { $ref: '#/components/responses/NotFound' }
  /runs/{run_id}/tournaments/{index}/hands:
    get:
      tags: [runs, admin]
      security: [{ AdminKey: [] }, { ApiKey: [] }]
      summary: Hand history (JSON lines) — admin only, only when record_hands was enabled
      parameters:
        - { name: run_id, in: path, required: true, schema: { type: string } }
        - { name: index, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: application/x-ndjson stream }
        "404": { $ref: '#/components/responses/NotFound' }
...